requirements.txt
streamlit>=1.32.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
langchain>=1.0.0
langchain-classic>=1.0.0
langchain-community>=0.4.0
langchain-core>=1.0.0
langchain-google-genai>=4.0.0
langchain-text-splitters>=1.0.0
PyPDF2>=3.0.0
faiss-cpu>=1.8.0